Fix computer player algorithm
Previously, after a strike, the computer player would try to hit
one random neighbour. If that neighbour hit didn't result in a hit,
the algorithm would then go on to hit either a predetermined position
or a completely random position.

Now, the app adds all neighbour position to the 'nextHit' array and
then makes sure that as long as the array isn't empty, it will fire
at a random position from that array.